+++ CHANGELOG.md
@@ -76,7 +76,7 @@ All notable changes to this project will be documented in this file.
   - `toStrictlyPositiveIntOrNull` and `toStrictlyPositiveIntOrThrow`
   - `toStrictlyNegativeIntOrNull` and `toStrictlyNegativeIntOrThrow`
   - `toStrictlyPositiveDoubleOrNull` and `toStrictlyPositiveDoubleOrThrow`
-  - `toNotBlankStringOrNull`.
+  - `toNotBlankStringOrNull` and `toNotBlankStringOrThrow`.
 
 [#258]: https://github.com/kotools/types/issues/258